Refactor DropdownField to abstract class and add RadioField implementation
